+++ docs/nuxt.config.ts
@@ -1,6 +1,6 @@
 // https://nuxt.com/docs/api/configuration/nuxt-config
 export default defineNuxtConfig({
-  modules: ['@nuxt/content', '@unocss/nuxt', "@nuxt/image"],
+  modules: ['@nuxt/content', '@unocss/nuxt', '@nuxt/image'],
   devtools: { enabled: true },
   app: {
     head: {
@@ -11,6 +11,11 @@ export default defineNuxtConfig({
     highlight: {
       theme: 'min-light',
       preload: ['ts', 'js', 'vue', 'bash'],
+    },
+  },
+  vite: {
+    define: {
+      __WEB__: 'true',
     },
   },
 })